+++ modules/culling/test/index.ts
@@ -12,6 +12,7 @@ import './lib/bounding-volumes/bounding-sphere.spec';
 import './lib/bounding-volumes/oriented-bounding-box.spec';
 
 import './lib/plane.spec';
+import './lib/ray.spec';
 import './lib/culling-volume.spec';
 
 import './lib/perspective-off-center-frustum.spec';